refactor: rename ClusterAdminSecurityConfiguration to ClusterAdminBasicSecurityConfiguration
Aligns the Basic chain's class name with its OIDC sibling
(ClusterAdminOidcSecurityConfiguration), per review feedback.
